Add files via upload
A mod for Hypixel Skyblock. 
Including an all-in-one package of Quality of Life upgrades for the beloved game, this mod includes various features for all aspects of the game such as:
Dungeon features - a 100% accurate puzzle solver that automatically turns off and on during the run, with improved access to them with the help of a wonderful sleek system that rivals all competitors, terminal helpers that reduce the risk of errors, secret auras help find areas that locate not just the secret, but the type and location of the entrance, a dungeon minimap that has a nodule for increases in scale factor and colour etc. the list is endless
General features - toggleable reparty features, that function with the new Crimson Isle update, a Bazaar tracker that notifies you in-game during a price crash, an auction house/buy-it-now sniper that functions quicker than all else, to be continued

Note: the mod does not function with Skytils